refactor(shared): split TUI agent config types into their own module
Adding IBM Bob pushed src/shared/tui-agent-config.ts to 303 effective
lines, over the 300-line max-lines cap for .ts files. AGENTS.md forbids
a max-lines suppression, so move the four type declarations into
tui-agent-config-types.ts, matching the existing *-types.ts convention
in src/shared.

Leaves the config module at 272 lines, so the next agent addition has
headroom rather than hitting the same wall.
